Move config defined recipes to a class with a better name.
