+++ src/Command/Recipe/ConfigDefinedRecipe.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace Waffle\Command;
+namespace Waffle\Command\Recipe;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
+use Waffle\Command\BaseCommand;
 
-class Recipe extends BaseCommand
+class ConfigDefinedRecipe extends BaseCommand
 {
     /**
      * @var string
@@ -23,7 +24,7 @@ class Recipe extends BaseCommand
         // TODO: Help and description are not properly set since these are
         // populated. Consider allow help and description text to be set in
         // config.
-        $help = 'Custom Recipe -- <comment>See Waffle config file.</comment>';
+        $help = 'Custom ConfigDefinedRecipe -- <comment>See Waffle config file.</comment>';
         $this->setDescription($help);
         $this->setHelp($help);
     }
@@ -73,7 +74,7 @@ class Recipe extends BaseCommand
 
             if ($return_code !== Command::SUCCESS) {
                 $this->io->highlightText(
-                    '[Recipe %s] Failed while running task %s',
+                    '[ConfigDefinedRecipe %s] Failed while running task %s',
                     [$this->config_key, $task_key],
                     'error',
                     'none'
@@ -83,7 +84,7 @@ class Recipe extends BaseCommand
             }
         }
 
-        $this->io->highlightText('Recipe %s complete!', [$this->config_key]);
+        $this->io->highlightText('ConfigDefinedRecipe %s complete!', [$this->config_key]);
 
         return Command::SUCCESS;
     }
